requirements.txt
absl-py==2.4.0
altair==5.2.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2026.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2025.9.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.7.1
mdurl==0.1.2
mediapipe==0.10.7
ml-dtypes==0.2.0
mpmath==1.3.0
narwhals==2.17.0
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
plotly==6.5.2
protobuf==3.20.3
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
referencing==0.37.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.4.2
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
streamlit==1.31.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
termcolor==2.4.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==1.1.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.1.0
validators==0.35.0
watchdog==3.0.0
Werkzeug==3.1.6
wrapt==1.14.1
zipp==3.17.0
